Chore[COMP345-78]: refactor character
