+++ GUI/ItemEditor.h
@@ -18,7 +18,7 @@
 #include <FL/Fl_Float_Input.H>
 
 #include "../ProjectSourceCode/item.h"
-#include "../ProjectSourceCode/Character.h"
+#include "../ProjectSourceCode/Character/Character.h"
 #include "../ProjectSourceCode/Serialize/serializeItem.h"
 
 #include "BaseEditor.h"
