+++ GUI/MapEditor.cpp
@@ -9,7 +9,7 @@
 #include "../ProjectSourceCode/Interactable/Interactable.h"
 #include "../ProjectSourceCode/Interactable/Wall.h"
 #include "../ProjectSourceCode/Interactable/EmptyCell.h"
-#include "../ProjectSourceCode/Character.h"
+#include "../ProjectSourceCode/Character/Character.h"
 #include "../ProjectSourceCode/item.h"
 #include "../ProjectSourceCode/Builder/MapBuilder.h"
