+++ ProjectSourceCode/Strategy/friendlystrategy.cpp
@@ -5,7 +5,7 @@
 */
 
 #include "friendlystrategy.h"
-#include "..\Character.h"
+#include "..\Character\Character.h"
 #include "..\Interactable\Wall.h"
 #include "..\Interactable\EmptyCell.h"
