+++ ProjectSourceCode/Strategy/humanplayerstrategy.cpp
@@ -5,7 +5,7 @@
 */
 
 #include "humanplayerstrategy.h"
-#include "..\Character.h"
+#include "..\Character\Character.h"
 #include "..\Interactable\Wall.h"
 #include "..\Interactable\EmptyCell.h"
