+++ ProjectSourceCode/Test/testCharacter.h
@@ -14,7 +14,7 @@
 #include <cppunit/CompilerOutputter.h>
 #include <cppunit/XmlOutputter.h>
 
-#include "..\Character.h"
+#include "..\Character\Character.h"
 
 class TestCharacter : public CppUnit::TestFixture
 {
