+++ [COMP345] Assignment 1.vcxproj.filters
@@ -27,7 +27,7 @@
     <ClCompile Include="ProjectSourceCode\Test\testitem.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ProjectSourceCode\Character.cpp">
+    <ClCompile Include="ProjectSourceCode\Character\Character.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="ProjectSourceCode\Dice\Dice.cpp">
@@ -143,7 +143,7 @@
     <ClInclude Include="ProjectSourceCode\Dice\Dice.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="ProjectSourceCode\Character.h">
+    <ClInclude Include="ProjectSourceCode\Character\Character.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="ProjectSourceCode\Test\testitemcontainer.h">
